Refactor. tooling traits to module Refactor. cleaning up cloud-init tooling
