+++ rs/src/tooling/az_cli.rs
@@ -1,5 +1,4 @@
-use crate::{Tooling, Tool};
-
+use super::Tooling;
 
 /// Built in az cli tool 
 pub struct AzCli;
